Move Aeon holdings logic into new Aeon helper file; refactor part of holdings HTML generation
