+++ blacklight-cornell/app/views/layouts/aeon/_holdings.html.erb
@@ -1,1 +1,2 @@
-<div id="holdings" class="scrollable"><%= @ho.html_safe %></div>
+<% aeon_holdings = holdings(@aeon_request.holdings, @aeon_request.items) %>
+<div id="holdings" class="scrollable"><%= aeon_holdings.html_safe %></div>
